fixing casting warnings

+++ kernel/src/vmm.c
@@ -123,7 +123,7 @@ void vmm_init()
     extern uint8_t *bitmap;
     extern uint64_t total_blocks;
 
-    vmm_maprange(kernel_pml4, bitmap, bitmap - hhdm_offset, total_blocks, PAGE_PRESENT | PAGE_WRITABLE);
+    vmm_maprange(kernel_pml4, (uint64_t)bitmap, (uint64_t)bitmap - hhdm_offset, total_blocks, PAGE_PRESENT | PAGE_WRITABLE);
 
     vmm_load_pml4();
 }
